+++ src/client/Main.ts
@@ -73,6 +73,7 @@ import "./components/DesktopNavBar";
 import "./components/Footer";
 import "./components/MainLayout";
 import "./components/MobileNavBar";
+import "./components/MobileTopBar";
 import "./components/PlayPage";
 import "./components/RankedModal";
 import "./components/baseComponents/Button";